requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Neo4j
neo4j>=5.17.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0
httpx>=0.26.0
aiohttp>=3.9.0

# XML parsing (for OFAC SDN)
lxml>=5.0.0
defusedxml>=0.7.0

# LLM integration
openai>=1.0.0
anthropic>=0.18.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
